Upgrade stats to uint64_t and replace EMA with ADAM bias correction
- Change all statistics counters from unsigned to uint64_t to prevent
  overflow after ~4.3B propagations on hard instances (CaDiCaL uses
  int64_t throughout).
- Replace halving-beta EMA initialization with ADAM-style bias
  correction (Kingma & Ba, ICLR 2015): value = biased / (1 - beta^t).
  Gives accurate estimates from the first update, improving early
  restart decisions.

+++ src/api/api_stats.cpp
@@ -109,7 +109,7 @@ extern "C" {
             SET_ERROR_CODE(Z3_INVALID_ARG, nullptr);
             return 0;
         }
-        return to_stats_ref(s).get_uint_value(idx);
+        return static_cast<unsigned>(to_stats_ref(s).get_uint_value(idx));
         Z3_CATCH_RETURN(0);
     }
